+++ api/src/main/java/org/testifyproject/annotation/Name.java
@@ -25,8 +25,7 @@ import java.lang.annotation.Target;
 
 /**
  * An annotation that can be placed on field, method, or method parameter to
- * associate it with a name. This allows you to distinguish fields, methods, and
- * method parameters using a custom name.
+ * associate it with a name.
  *
  * @author saden
  */