+++ junit4/junit4-core/src/test/java/org/testifyproject/junit4/ExplicitNameIndistinctGenericTypeTest.java
@@ -24,8 +24,9 @@ import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
-import org.testifyproject.annotation.Sut;
 import org.testifyproject.annotation.Fake;
+import org.testifyproject.annotation.Name;
+import org.testifyproject.annotation.Sut;
 import org.testifyproject.junit4.fixture.ExplicitNameIndistinctGenericType;
 import org.testifyproject.junit4.fixture.common.Hello;
 
@@ -42,7 +43,8 @@ public class ExplicitNameIndistinctGenericTypeTest {
     @Fake
     Supplier<Hello> english;
 
-    @Fake("spanish")
+    @Name("spanish")
+    @Fake
     Supplier<Hello> esp;
 
     @Before